[feat] 선택 마커 주소데이터 컴포넌트 분리  #43
