+++ src/components/Main/ApartMap/ApartLocation.module.scss
@@ -1,0 +1,22 @@
+@use "@/styles/colors.scss";
+
+.location {
+  position: absolute;
+  display: flex;
+  z-index: 50;
+  align-items: center;
+  gap: 6rem;
+  background-color: colors.$mainTitle;
+  padding: 1.2rem 2.4rem;
+  top: 0;
+  left: 0;
+  cursor: pointer;
+  span {
+    color: #ffffff;
+    font-size: 2rem;
+    font-weight: 600;
+  }
+  img {
+    width: 2rem;
+  }
+}
