+++ src/components/Main/ApartMap/ApartMap.tsx
@@ -5,22 +5,17 @@ import styles from "./ApartMap.module.scss";
 import { useApartInfoStore } from "@/store/useApartInfoStore";
 import DetailInfo from "../DetailInfo/DetailInfo";
 import MapLoading from "./MapLoading";
-import locationIcon from "@/assets/Main/Map/locationIcon.svg";
 import CurrentMapSearch from "./CurrentMapSearch";
 import useGetApartMarker from "@/hooks/Api/useGetApartMarker";
 import { useMarkerStore } from "@/store/useMarkerStore";
 import useMapMarkers from "@/hooks/Map/useMapMarkers";
 import useCreateMap from "@/hooks/Map/useCreateMap";
 import useLocationPath from "@/hooks/Map/useLocationPath";
-import { useWeakApartInfoStore } from "@/store/useWeakApartInfoStore";
-import { useMainInfoStore } from "@/store/useMainInfoStore";
 import { updateBounds } from "@/utils/map/updateBounds";
+import ApartLocation from "./ApartLocation";
 
 export default function ApartMap() {
   const isDetailInfo = useApartInfoStore((state) => state.isDetailInfo);
-  const weakApartInfo = useWeakApartInfoStore((state) => state.weakApartInfo);
-  const apartInfo = useMainInfoStore((state) => state.apartInfo);
-  const selectMarker = useMarkerStore((state) => state.selectMarker);
   const map = useMarkerStore((state) => state.map);
   const setMap = useMarkerStore((state) => state.setMap);
   const bounds = useMarkerStore((state) => state.bounds);
@@ -31,16 +26,6 @@ export default function ApartMap() {
   const updateBoundsHandler = () => {
     if (map) {
       updateBounds(map, setBounds);
-    }
-  };
-
-  const locationClickHandler = () => {
-    if (selectMarker && map) {
-      const newLocation = new naver.maps.LatLng(
-        selectMarker?.getPosition().y,
-        selectMarker?.getPosition().x
-      );
-      map.panTo(newLocation);
     }
   };
 
@@ -60,18 +45,10 @@ export default function ApartMap() {
 
   useMapMarkers();
 
+  console.log(mapRef);
   return (
     <section className={styles.apartMap}>
-      {!isLoading && (
-        <div className={styles.location} onClick={locationClickHandler}>
-          <span>
-            {locationPath === "apt"
-              ? apartInfo?.aptInfo.roadAddress
-              : weakApartInfo?.aptInfo.address}
-          </span>
-          <img src={locationIcon} alt="map location" />
-        </div>
-      )}
+      {!isLoading && <ApartLocation />}
       {isDetailInfo && <DetailInfo />}
       {isLoading ? (
         <MapLoading />
